+++ mrv2/lib/mrvFl/mrvSession.h
@@ -5,16 +5,26 @@
 #pragma once
 
 #include <string>
+#include <map>
 
 namespace mrv
 {
     namespace session
     {
-        //! Returns user provided metadata in the last saved session.
-        const std::string& metadata();
+        //! Returns all metadata.
+        const std::map<std::string, std::string>& metadata();
+
+        //! Given a key returns user provided metadata in the current session.
+        const std::string& metadata(const std::string& key);
 
         //! Stores some user provided metadata.
-        void setMetadata(const std::string& metadata);
+        void setMetadata(const std::string& key, const std::string& value);
+
+        //! Replaces all metadata.
+        void setMetadata(const std::map<std::string, std::string>&);
+
+        //! Clears the metadata for the current session.
+        void clearMetadata();
 
         //! Returns the current session file name.
         std::string current();